requirements.txt
markitdown
python-docx
openpyxl
python-pptx
pandas
chardet
py7zr
rarfile
lhafile
python-magic-bin
Pillow
tqdm
pyyaml
